+++ src/features/lobby/index.ts
@@ -5,3 +5,4 @@ export { TeamSection } from './ui/TeamSection';
 export { TeamTab } from './ui/TeamTab';
 export { useRoomUI } from './ui/useRoomUI';
 export { NudgeModal } from './ui/NudgeModal';
+export { useNudgeListener } from './model/useNudgeListener';
