+++ src/features/shop/index.ts
@@ -3,4 +3,5 @@ export * from './model/types';
 export { useShopFilter } from './model/useShopFilter';
 export { useShopPreview } from './model/useShopPreview';
 export { useProductSorting } from './model/useProductSorting';
+export { useShop } from './model/useShop';
 export { shopQueries } from './queries/shopQueries';
